requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pymupdf
pdf2image
Pillow
pytesseract
mrz
spacy
dateparser
phonenumbers
pydantic
email-validator
openai
playwright
python-dotenv
aiofiles
